refactor(tics): converted v3 components directories to modules
