+++ src/app/pools/components/index.ts
@@ -1,0 +1,5 @@
+export { default as AddPool } from "./AddPool";
+export { default as DeletePool } from "./DeletePool";
+export { default as EditPool } from "./EditPool";
+export { default as PoolsListHeader } from "./PoolsListHeader";
+export { default as PoolsTable } from "./PoolsTable";
