+++ src/app/preferences/views/SSHKeys/components/SSHKeysTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SSHKeysTable";
