+++ src/app/preferences/views/SSHKeys/components/index.ts
@@ -1,0 +1,3 @@
+export { default as AddSSHKey } from "./AddSSHKey";
+export { default as DeleteSSHKey } from "./DeleteSSHKey";
+export { default as SSHKeysTable } from "./SSHKeysTable";
