+++ src/app/preferences/views/SSLKeys/components/SSLKeysTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SSLKeysTable";
