+++ src/app/preferences/views/SSLKeys/components/index.ts
@@ -1,0 +1,3 @@
+export { default as AddSSLKey } from "./AddSSLKey";
+export { default as DeleteSSLKey } from "./DeleteSSLKey";
+export { default as SSLKeysTable } from "./SSLKeysTable";
