+++ src/app/settings/views/Users/components/AddUser/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddUser";
