+++ src/app/settings/views/Users/components/DeleteUser/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DeleteUser";
