+++ src/app/settings/views/Users/components/EditUser/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./EditUser";
