+++ src/app/settings/views/Users/components/index.ts
@@ -1,0 +1,4 @@
+export { default as AddUser } from "./AddUser";
+export { default as DeleteUser } from "./DeleteUser";
+export { default as EditUser } from "./EditUser";
+export { default as UsersTable } from "./UsersTable";
