+++ src/app/zones/components/index.ts
@@ -1,0 +1,5 @@
+export { default as AddZone } from "./AddZone";
+export { default as DeleteZone } from "./DeleteZone";
+export { default as EditZone } from "./EditZone";
+export { default as ZonesListHeader } from "./ZonesListHeader";
+export { default as ZonesTable } from "./ZonesTable";
